app/requirements.txt
Django>=5.1.1
djangorestframework>=3.15,<3.16
# flake8
psycopg2>=2.8.1,<3.0
drf-spectacular>=0.15.1,<0.16.0
httpcore>=0.16.0,<0.17.0